show_plant_biomes is modular now.
