+++ actions/__init__.py
@@ -4,3 +4,4 @@ from .report import build_facility_report
 from .feed_animal import feed_animal
 from .cultivate_plant import cultivate_plant
 from .get_plant_biomes import get_plant_biomes
+from .show_plant_biomes import show_plant_biomes
